requirements.txt
aiohttp==3.9.5
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
fpdf==1.7.2
together==1.2.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
lxml==5.3.1
multidict==6.2.0
numpy==2.2.4
pandas==2.2.3
pillow==10.3.0
propcache==0.3.1
pydub==0.25.1
python-dateutil==2.9.0.post0
python-telegram-bot==20.7
pytrends==4.9.2
pytz==2025.2
reportlab==4.3.1
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.10.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
yarl==1.18.3